+++ flake.nix
@@ -20,7 +20,7 @@
 
         himitsu-bako = pkgs.buildGoModule {
           pname = "himitsu-bako";
-          version = "1.0.0";
+          version = "1.1.0";
 
           src = self;